Project/Django-workspace/requirements.txt
requests
django==2.2
djangorestframework==3.9.2